@@ -13,33 +13,34 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/common": "9.1.6",
-        "@angular/core": "9.1.6",
-        "@angular/forms": "9.1.6",
-        "@angular/platform-browser": "9.1.6",
-        "@angular/platform-browser-dynamic": "9.1.6",
-        "@angular/router": "9.1.6",
+        "@angular-devkit/core": "^16.2.1",
+        "@angular/common": "16.2.4",
+        "@angular/core": "16.2.4",
+        "@angular/forms": "16.2.4",
+        "@angular/platform-browser": "16.2.4",
+        "@angular/platform-browser-dynamic": "16.2.4",
+        "@angular/router": "16.2.4",
         "@ionic-native/core": "5.36.0",
         "@ionic-native/splash-screen": "5.36.0",
         "@ionic-native/status-bar": "5.36.0",
-        "@ionic/angular": "5.0.0",
-        "cordova-android": "9.1.0",
-        "cordova-ios": "6.2.0",
+        "@ionic/angular": "^6.7.5",
+        "cordova-android": "^12.0.1",
+        "cordova-ios": "^7.0.1",
         "nan": "2.14.2",
-        "rxjs": "6.5.3",
-        "scandit-cordova-datacapture-barcode": "6.19.5",
-        "scandit-cordova-datacapture-core": "6.19.5",
-        "tslib": "1.10.0",
-        "zone.js": "0.10.3"
+        "rxjs": "^6.6.0",
+        "scandit-cordova-datacapture-barcode": "6.20.0",
+        "scandit-cordova-datacapture-core": "6.20.0",
+        "tslib": "^2.0.0",
+        "zone.js": "~0.13.2"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "0.901.5",
-        "@angular/cli": "9.1.5",
-        "@angular/compiler": "9.1.6",
-        "@angular/compiler-cli": "9.1.6",
-        "@angular/language-service": "9.1.6",
+        "@angular-devkit/build-angular": "16.2.1",
+        "@angular/cli": "16.2.1",
+        "@angular/compiler": "16.2.4",
+        "@angular/compiler-cli": "16.2.4",
+        "@angular/language-service": "16.2.4",
         "@ionic/angular-toolkit": "2.1.1",
-        "@types/jasmine": "3.5.0",
+        "@types/jasmine": "~3.6.0",
         "@types/jasminewd2": "2.0.3",
         "@types/node": "12.11.1",
         "cordova-plugin-device": "2.0.2",
@@ -48,8 +49,8 @@
         "cordova-plugin-splashscreen": "6.0.0",
         "cordova-plugin-statusbar": "2.4.2",
         "ts-node": "8.3.0",
-        "tslint": "6.1.0",
-        "typescript": "3.8.3"
+        "tslint": "~6.1.0",
+        "typescript": "4.9.3"
     },
     "cordova": {
         "plugins": {
