@@ -6,37 +6,37 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/common": "~13.0.0",
-        "@angular/core": "~13.0.0",
-        "@angular/forms": "~13.0.0",
-        "@angular/platform-browser": "~13.0.0",
-        "@angular/platform-browser-dynamic": "~13.0.0",
-        "@angular/router": "~13.0.0",
+        "@angular/common": "^16.2.5",
+        "@angular/core": "^16.2.5",
+        "@angular/forms": "^16.2.5",
+        "@angular/platform-browser": "^16.2.5",
+        "@angular/platform-browser-dynamic": "^16.2.5",
+        "@angular/router": "^16.2.5",
         "@ionic-native/core": "5.36.0",
         "@ionic-native/splash-screen": "5.36.0",
         "@ionic-native/status-bar": "5.36.0",
         "@ionic/angular": "^6.0.0",
         "rxjs": "~6.6.0",
-        "scandit-cordova-datacapture-barcode": "6.19.5",
-        "scandit-cordova-datacapture-core": "6.19.5",
+        "scandit-cordova-datacapture-barcode": "6.20.0",
+        "scandit-cordova-datacapture-core": "6.20.0",
         "tslib": "^2.2.0",
-        "zone.js": "~0.11.4"
+        "zone.js": "~0.13.3"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~13.0.1",
+        "@angular-devkit/build-angular": "^16.2.2",
         "@angular-eslint/builder": "~13.0.1",
         "@angular-eslint/eslint-plugin": "~13.0.1",
         "@angular-eslint/eslint-plugin-template": "~13.0.1",
         "@angular-eslint/template-parser": "~13.0.1",
-        "@angular/cli": "~13.0.1",
-        "@angular/compiler": "~13.0.0",
-        "@angular/compiler-cli": "~13.0.0",
-        "@angular/language-service": "~13.0.0",
+        "@angular/cli": "^16.2.2",
+        "@angular/compiler": "^16.2.5",
+        "@angular/compiler-cli": "^16.2.5",
+        "@angular/language-service": "^16.2.5",
         "@ionic/angular-toolkit": "^5.0.0",
         "@types/node": "^12.11.1",
         "@typescript-eslint/eslint-plugin": "5.3.0",
         "@typescript-eslint/parser": "5.3.0",
-        "cordova-ios": "^6.2.0",
+        "cordova-ios": "^7.0.1",
         "cordova-plugin-device": "^2.0.2",
         "cordova-plugin-ionic-keyboard": "^2.2.0",
         "cordova-plugin-ionic-webview": "^5.0.0",
@@ -47,7 +47,7 @@
         "eslint-plugin-jsdoc": "30.7.6",
         "eslint-plugin-prefer-arrow": "1.2.2",
         "ts-node": "~8.3.0",
-        "typescript": "~4.4.4"
+        "typescript": "~4.9.3"
     },
     "cordova": {
         "plugins": {
